move everything in game/play into game/
