+++ dp-game-webapp/app/game/results.tsx
@@ -7,7 +7,7 @@ import {
 } from "@heroicons/react/24/outline";
 
 import { Answer, AnsweredQuestion } from "./questions";
-import { ExponentialNumber } from "../../exponentialNumber";
+import { ExponentialNumber } from "../exponentialNumber";
 import { GameContainer, PageContainer, PageTitle } from "./components";
 
 const answerIcons: Record<Answer, ReactNode> = {
