+++ dp-game-webapp/app/game/validate.tsx
@@ -5,14 +5,14 @@ import {
   simulatedPercentiles,
   increaseVariance,
   decreaseVariance,
-} from "../simulate";
+} from "./simulate";
 import {
   ArrowRightCircleIcon,
   ArrowLeftCircleIcon,
 } from "@heroicons/react/24/outline";
 
 import AdjustVariance from "./adjustVariance";
-import { ExponentialNumber } from "../../exponentialNumber";
+import { ExponentialNumber } from "../exponentialNumber";
 import {
   GameContainer,
   PageContainer,
